Change all files to Unix line endings
This will make working with and patching the code more pleasant. This
was done automatically with the following command:

    find . -type f -not -ipath './.git*' \
        -not -name '*.bmp' -not -name '*.png' -not -name '*.bin' \
        -exec sh -c 'file {} | grep CRLF' \; \
        -exec sed -i 's/\r//g;s/\s*$//' {} \;

However, this is likely to break all PRs that have been created before
this change is merged.

As a way to work around this issue, the command above can be used on the
tip of the PR to convert the modified files to the same line endings.
The PR branch can then be recreated the new master (incorporating this
commit), and the changed files copied/committed over again. A final

    git push --force-with-lease

will set the PR straight and back to mergeable (at least as far as
whitespaces are concerned).

Signed-off-by: Olivier Mehani <shtrom@ssji.net>